#796: Extract liquidity pool storage types into their own module
lib.rs mixed the Soroban contract implementation with its storage key
enum and data structs. Move DataKey, PoolConfig, SwapRecord,
MiningConfig, and PoolStats into types.rs for a clearer boundary
between data definitions and contract logic.
